move out backtrace decoding
