+++ website/agb/src/app/crash/page.tsx
@@ -1,6 +1,5 @@
 import { Metadata } from "next";
 import { BacktracePage } from "./backtrace";
-import { ContentBlock } from "../contentBlock";
 
 export const metadata: Metadata = {
   title: "agbrs crash backtrace",
